clean build_dir only if it exists

+++ noxfile.py
@@ -200,9 +200,10 @@ def clean(session):
 @nox.session(python=False, name="clean-docs")
 def clean_docs(session: nox.Session) -> None:
     """Clean up the docs folder."""
-    with session.chdir(ROOT / "build"):
-        if os.path.exists("html"):
-            shutil.rmtree("html")
+    if (ROOT / "build").is_dir():
+        with session.chdir(ROOT / "build"):
+            if os.path.exists("html"):
+                shutil.rmtree("html")
 
     with session.chdir(ROOT / "docs"):
         for p in pathlib.Path("api").rglob("sequence*.rst"):
